perf: show mini stats widget on admin dashboard, with link to full widget
